+++ tests/test_websocket.py
@@ -151,3 +151,18 @@ def test_disallow_body_inputs():
 
     with pytest.raises(InvalidRouteError):
         app.load()
+
+@limit_leaks("1 MB")
+@pytest.mark.asyncio
+async def test_websocket_leaks():
+    app = new_app()
+
+    @app.websocket("/")
+    async def index(ws: WebSocket):
+        await ws.accept()
+        await ws.close()
+
+    async with app.test() as test:
+        for i in range(10000):
+            async with test.websocket("/") as ws:
+                ...